체력바 추가, 데미지폰트 추가, 데미지 프로세스 수정
체력바 추가, 데미지폰트 추가, 데미지 프로세스 수정

+++ Project/Assets/Main/Scripts/FrameWork/Controller/Managers/Player/PlayerManager.cs
@@ -47,7 +47,7 @@ namespace GameBerry.Managers
             return m_playerController;
         }
         //------------------------------------------------------------------------------------
-        public int GetCurrentPlayerMP()
+        public double GetCurrentPlayerMP()
         {
             return m_playerController.CurrentMP;
         }
